+++ gazebo/gui/MainWindow.cc
@@ -238,6 +238,10 @@ MainWindow::MainWindow()
   connect(this, SIGNAL(AddPlugins()),
           this, SLOT(OnAddPlugins()), Qt::QueuedConnection);
 
+  // Use a signal/slot to track a visual. This makes the process thread safe.
+  connect(this, SIGNAL(TrackVisual(const std::string &)),
+          this, SLOT(OnTrackVisual(const std::string &)), Qt::QueuedConnection);
+
   // Create data logger dialog
   this->dataPtr->dataLogger = new gui::DataLogger(this);
   connect(this->dataPtr->dataLogger, SIGNAL(rejected()), this, SLOT(
@@ -1953,17 +1957,38 @@ void MainWindow::OnGUI(ConstGUIPtr &_msg)
 
     if (_msg->camera().has_track())
     {
-      std::string name = _msg->camera().track().name();
+      if (_msg->camera().track().has_static_())
+        cam->SetTrackIsStatic(_msg->camera().track().static_());
 
-      double minDist = 0.0;
-      double maxDist = 0.0;
+      if (_msg->camera().track().has_use_model_frame())
+        cam->SetTrackUseModelFrame(_msg->camera().track().use_model_frame());
+
+      if (_msg->camera().track().has_xyz())
+        cam->SetTrackPosition(msgs::ConvertIgn(_msg->camera().track().xyz()));
+
+      if (_msg->camera().track().has_inherit_yaw())
+        cam->SetTrackInheritYaw(_msg->camera().track().inherit_yaw());
 
       if (_msg->camera().track().has_min_dist())
-        minDist = _msg->camera().track().min_dist();
-      if (_msg->camera().track().has_max_dist())
-        maxDist = _msg->camera().track().max_dist();
+      {
+        double minDist = _msg->camera().track().min_dist();
+        cam->SetTrackMinDistance(minDist);
+      }
 
-      cam->AttachToVisual(name, false, minDist, maxDist);
+      if (_msg->camera().track().has_max_dist())
+      {
+        double maxDist = _msg->camera().track().max_dist();
+        cam->SetTrackMaxDistance(maxDist);
+      }
+
+      if (_msg->camera().track().has_name() &&
+          _msg->camera().track().name() != "__default__")
+      {
+        std::string name = _msg->camera().track().name();
+        cam->TrackVisual(name);
+        // Call the signal to track a visual in the main thread.
+        this->TrackVisual(name);
+      }
     }
   }
 
@@ -2016,6 +2041,12 @@ void MainWindow::OnAddPlugins()
 
   g_overlayAct->setChecked(true);
   g_overlayAct->setEnabled(true);
+}
+
+/////////////////////////////////////////////////
+void MainWindow::OnTrackVisual(const std::string &_visualName)
+{
+  gui::Events::follow(_visualName);
 }
 
 /////////////////////////////////////////////////