+++ gazebo/gui/model/ModelData_TEST.cc
@@ -195,7 +195,7 @@ void ModelData_TEST::LinkScale()
     msgs::AddBoxLink(model, mass, size);
     link->Load(msgs::LinkToSDF(model.link(0)));
     rendering::VisualPtr linkVis(new rendering::Visual("box_link", scene));
-    link->linkVisual = linkVis;
+    link->SetLinkVisual(linkVis);
 
     // add a collision visual
     rendering::VisualPtr collisionVis(
@@ -344,7 +344,7 @@ void ModelData_TEST::LinkScale()
     msgs::AddCylinderLink(model, mass, radius, length);
     link->Load(msgs::LinkToSDF(model.link(0)));
     rendering::VisualPtr linkVis(new rendering::Visual("cylinder_link", scene));
-    link->linkVisual = linkVis;
+    link->SetLinkVisual(linkVis);
 
     // add a collision visual
     rendering::VisualPtr collisionVis(
@@ -498,7 +498,7 @@ void ModelData_TEST::LinkScale()
     msgs::AddSphereLink(model, mass, radius);
     link->Load(msgs::LinkToSDF(model.link(0)));
     rendering::VisualPtr linkVis(new rendering::Visual("sphere_link", scene));
-    link->linkVisual = linkVis;
+    link->SetLinkVisual(linkVis);
 
     // add a collision visual
     rendering::VisualPtr collisionVis(
@@ -646,7 +646,7 @@ void ModelData_TEST::LinkScale()
     msgs::AddSphereLink(model, mass, radius);
     link->Load(msgs::LinkToSDF(model.link(0)));
     rendering::VisualPtr linkVis(new rendering::Visual("sphere_link2", scene));
-    link->linkVisual = linkVis;
+    link->SetLinkVisual(linkVis);
 
     // add a collision visual
     rendering::VisualPtr collisionVis(
@@ -746,7 +746,7 @@ void ModelData_TEST::LinkScale()
     msgs::AddSphereLink(model, mass, radius);
     link->Load(msgs::LinkToSDF(model.link(0)));
     rendering::VisualPtr linkVis(new rendering::Visual("sphere_link3", scene));
-    link->linkVisual = linkVis;
+    link->SetLinkVisual(linkVis);
 
     // add a collision visual
     rendering::VisualPtr collisionVis(
@@ -824,6 +824,394 @@ void ModelData_TEST::LinkScale()
   mainWindow->close();
   delete mainWindow;
   mainWindow = NULL;
+}
+
+/////////////////////////////////////////////////
+void ModelData_TEST::LinkVolume()
+{
+  this->resMaxPercentChange = 5.0;
+  this->shareMaxPercentChange = 2.0;
+
+  this->Load("worlds/empty.world");
+
+  gazebo::gui::MainWindow *mainWindow = new gazebo::gui::MainWindow();
+  QVERIFY(mainWindow != NULL);
+
+  mainWindow->Load();
+  mainWindow->Init();
+  mainWindow->show();
+
+  gazebo::rendering::UserCameraPtr cam = gazebo::gui::get_active_camera();
+  QVERIFY(cam != NULL);
+
+  gazebo::rendering::ScenePtr scene = cam->GetScene();
+  QVERIFY(scene != NULL);
+
+  for (int i = 0; i < 10; ++i)
+  {
+    gazebo::common::Time::MSleep(30);
+    QCoreApplication::processEvents();
+    mainWindow->repaint();
+  }
+
+  //  Verify box volume calcs
+  {
+    gui::LinkData *link = new gui::LinkData();
+
+    ignition::math::Vector3d size(3, 4, 5);
+    double mass = 1.0;
+    double expectedVolume = 60.0;
+
+    msgs::Model model;
+    msgs::AddBoxLink(model, mass, size);
+
+    link->Load(msgs::LinkToSDF(model.link(0)));
+
+    rendering::VisualPtr linkVis(new rendering::Visual("box_link", scene));
+    link->SetLinkVisual(linkVis);
+
+    rendering::VisualPtr collisionVis(
+        new rendering::Visual("box_link::collision", linkVis));
+
+    collisionVis->Load(msgs::VisualToSDF(model.link(0).visual(0)));
+    link->AddCollision(collisionVis);
+
+    double volume = link->ComputeVolume();
+    QVERIFY(fabs(expectedVolume - volume) < 1e-3);
+
+    delete link;
+  }
+
+  // Verify sphere volume calcs
+  {
+    gui::LinkData *link = new gui::LinkData();
+
+    double mass = 1.0;
+    double radius = 1.5;
+    double expectedVolume = 14.137;
+
+    msgs::Model model;
+    msgs::AddSphereLink(model, mass, radius);
+
+    link->Load(msgs::LinkToSDF(model.link(0)));
+
+    rendering::VisualPtr linkVis(new rendering::Visual("sphere_link", scene));
+    link->SetLinkVisual(linkVis);
+
+    rendering::VisualPtr collisionVis(
+        new rendering::Visual("sphere_link::collision", linkVis));
+
+    collisionVis->Load(msgs::VisualToSDF(model.link(0).visual(0)));
+    link->AddCollision(collisionVis);
+
+    double volume = link->ComputeVolume();
+    QVERIFY(fabs(expectedVolume - volume) < 1e-3);
+
+    delete link;
+  }
+
+  // Verify cylinder volume calcs
+  {
+    gui::LinkData *link = new gui::LinkData();
+
+    double mass = 1.0;
+    double radius = 1.5;
+    double length = 5.0;
+    double expectedVolume = 35.343;
+
+    msgs::Model model;
+    msgs::AddCylinderLink(model, mass, radius, length);
+
+    link->Load(msgs::LinkToSDF(model.link(0)));
+
+    rendering::VisualPtr linkVis(new rendering::Visual("cylinder_link", scene));
+    link->SetLinkVisual(linkVis);
+
+    rendering::VisualPtr collisionVis(
+        new rendering::Visual("cylinder_link::collision", linkVis));
+
+    collisionVis->Load(msgs::VisualToSDF(model.link(0).visual(0)));
+    link->AddCollision(collisionVis);
+
+    double volume = link->ComputeVolume();
+    QVERIFY(fabs(expectedVolume - volume) < 1e-3);
+
+    delete link;
+  }
+}
+
+/////////////////////////////////////////////////
+void ModelData_TEST::BoxVolume()
+{
+  const double l = 3, w = 4, h = 5;
+  const double expectedVolume = 60;
+
+  msgs::Vector3d *size = new msgs::Vector3d();
+  size->set_x(l);
+  size->set_y(w);
+  size->set_z(h);
+
+  msgs::BoxGeom *box = new msgs::BoxGeom();
+  box->set_allocated_size(size);
+
+  msgs::Geometry *geo = new msgs::Geometry();
+  geo->set_type(msgs::Geometry_Type_BOX);
+  geo->set_allocated_box(box);
+
+  msgs::Collision *col = new msgs::Collision();
+  col->set_allocated_geometry(geo);
+
+  double volume = gui::LinkData::ComputeVolume(*col);
+  QVERIFY(fabs(expectedVolume - volume) < 1e-3);
+
+  delete col;
+}
+
+/////////////////////////////////////////////////
+void ModelData_TEST::CylinderVolume()
+{
+  const double r = 1.5, l = 5.0;
+  const double expectedVolume = 35.343;
+
+  msgs::CylinderGeom *cyl = new msgs::CylinderGeom();
+  cyl->set_radius(r);
+  cyl->set_length(l);
+
+  msgs::Geometry *geo = new msgs::Geometry();
+  geo->set_type(msgs::Geometry_Type_CYLINDER);
+  geo->set_allocated_cylinder(cyl);
+
+  msgs::Collision *col = new msgs::Collision();
+  col->set_allocated_geometry(geo);
+
+  double volume = gui::LinkData::ComputeVolume(*col);
+  QVERIFY(fabs(expectedVolume - volume) < 1e-3);
+
+  delete col;
+}
+
+/////////////////////////////////////////////////
+void ModelData_TEST::SphereVolume()
+{
+  const double r = 1.5;
+  const double expectedVolume = 14.137;
+
+  msgs::SphereGeom *sphere = new msgs::SphereGeom();
+  sphere->set_radius(r);
+
+  msgs::Geometry *geo = new msgs::Geometry();
+  geo->set_type(msgs::Geometry_Type_SPHERE);
+  geo->set_allocated_sphere(sphere);
+
+  msgs::Collision *col = new msgs::Collision();
+  col->set_allocated_geometry(geo);
+
+  double volume = gui::LinkData::ComputeVolume(*col);
+  QVERIFY(fabs(expectedVolume - volume) < 1e-3);
+
+  delete col;
+}
+
+/////////////////////////////////////////////////
+void ModelData_TEST::MeshVolume()
+{
+  const double l = 3, w = 4, h = 5;
+  const double expectedVolume = 60;
+
+  msgs::Vector3d *size = new msgs::Vector3d();
+  size->set_x(l);
+  size->set_y(w);
+  size->set_z(h);
+
+  msgs::BoxGeom *box = new msgs::BoxGeom();
+  box->set_allocated_size(size);
+
+  msgs::Geometry *geo = new msgs::Geometry();
+  geo->set_type(msgs::Geometry_Type_MESH);
+  geo->set_allocated_box(box);
+
+  msgs::Collision *col = new msgs::Collision();
+  col->set_allocated_geometry(geo);
+
+  double volume = gui::LinkData::ComputeVolume(*col);
+  QVERIFY(fabs(expectedVolume - volume) < 1e-3);
+
+  delete col;
+}
+
+/////////////////////////////////////////////////
+void ModelData_TEST::PolylineVolume()
+{
+  const double l = 3, w = 4, h = 5;
+  const double expectedVolume = 60;
+
+  msgs::Vector3d *size = new msgs::Vector3d();
+  size->set_x(l);
+  size->set_y(w);
+  size->set_z(h);
+
+  msgs::BoxGeom *box = new msgs::BoxGeom();
+  box->set_allocated_size(size);
+
+  msgs::Geometry *geo = new msgs::Geometry();
+  geo->set_type(msgs::Geometry_Type_POLYLINE);
+  geo->set_allocated_box(box);
+
+  msgs::Collision *col = new msgs::Collision();
+  col->set_allocated_geometry(geo);
+
+  double volume = gui::LinkData::ComputeVolume(*col);
+  QVERIFY(fabs(expectedVolume - volume) < 1e-3);
+
+  delete col;
+}
+
+void ModelData_TEST::SphereMomentOfInertia()
+{
+  const double r = 1.5;
+  const double m = 1.0;
+  const double expectedI = 0.9;
+
+  msgs::SphereGeom *sphere = new msgs::SphereGeom();
+  sphere->set_radius(r);
+
+  msgs::Geometry *geo = new msgs::Geometry();
+  geo->set_type(msgs::Geometry_Type_SPHERE);
+  geo->set_allocated_sphere(sphere);
+
+  msgs::Collision *col = new msgs::Collision();
+  col->set_allocated_geometry(geo);
+
+  ignition::math::Vector3d I = gui::LinkData::ComputeMomentOfInertia(*col, m);
+  QVERIFY(fabs(expectedI - I.X()) < 1e-3);
+  QVERIFY(fabs(expectedI - I.Y()) < 1e-3);
+  QVERIFY(fabs(expectedI - I.Z()) < 1e-3);
+
+  delete col;
+}
+
+/////////////////////////////////////////////////
+void ModelData_TEST::CylinderMomentOfInertia()
+{
+  const double r = 1.5, l = 5.0, m = 1.0;
+  const double expectedIx = 2.64583;
+  const double expectedIy = 2.64583;
+  const double expectedIz = 1.125;
+
+  msgs::CylinderGeom *cyl = new msgs::CylinderGeom();
+  cyl->set_radius(r);
+  cyl->set_length(l);
+
+  msgs::Geometry *geo = new msgs::Geometry();
+  geo->set_type(msgs::Geometry_Type_CYLINDER);
+  geo->set_allocated_cylinder(cyl);
+
+  msgs::Collision *col = new msgs::Collision();
+  col->set_allocated_geometry(geo);
+
+  ignition::math::Vector3d I = gui::LinkData::ComputeMomentOfInertia(*col, m);
+
+  QVERIFY(fabs(expectedIx - I.X()) < 1e-3);
+  QVERIFY(fabs(expectedIy - I.Y()) < 1e-3);
+  QVERIFY(fabs(expectedIz - I.Z()) < 1e-3);
+
+  delete col;
+}
+
+/////////////////////////////////////////////////
+void ModelData_TEST::BoxMomentOfInertia()
+{
+  const double l = 3, w = 4, h = 5, m = 1.0;
+  const double expectedIx = 3.41667;
+  const double expectedIy = 2.83333;
+  const double expectedIz = 2.08333;
+
+  msgs::Vector3d *size = new msgs::Vector3d();
+  size->set_x(l);
+  size->set_y(w);
+  size->set_z(h);
+
+  msgs::BoxGeom *box = new msgs::BoxGeom();
+  box->set_allocated_size(size);
+
+  msgs::Geometry *geo = new msgs::Geometry();
+  geo->set_type(msgs::Geometry_Type_BOX);
+  geo->set_allocated_box(box);
+
+  msgs::Collision *col = new msgs::Collision();
+  col->set_allocated_geometry(geo);
+
+  ignition::math::Vector3d I = gui::LinkData::ComputeMomentOfInertia(*col, m);
+
+  QVERIFY(fabs(expectedIx - I.X()) < 1e-3);
+  QVERIFY(fabs(expectedIy - I.Y()) < 1e-3);
+  QVERIFY(fabs(expectedIz - I.Z()) < 1e-3);
+
+  delete col;
+}
+
+/////////////////////////////////////////////////
+void ModelData_TEST::MeshMomentOfInertia()
+{
+  const double l = 3, w = 4, h = 5, m = 1.0;
+  const double expectedIx = 3.41667;
+  const double expectedIy = 2.83333;
+  const double expectedIz = 2.08333;
+
+  msgs::Vector3d *size = new msgs::Vector3d();
+  size->set_x(l);
+  size->set_y(w);
+  size->set_z(h);
+
+  msgs::BoxGeom *box = new msgs::BoxGeom();
+  box->set_allocated_size(size);
+
+  msgs::Geometry *geo = new msgs::Geometry();
+  geo->set_type(msgs::Geometry_Type_MESH);
+  geo->set_allocated_box(box);
+
+  msgs::Collision *col = new msgs::Collision();
+  col->set_allocated_geometry(geo);
+
+  ignition::math::Vector3d I = gui::LinkData::ComputeMomentOfInertia(*col, m);
+
+  QVERIFY(fabs(expectedIx - I.X()) < 1e-3);
+  QVERIFY(fabs(expectedIy - I.Y()) < 1e-3);
+  QVERIFY(fabs(expectedIz - I.Z()) < 1e-3);
+
+  delete col;
+}
+
+/////////////////////////////////////////////////
+void ModelData_TEST::PolylineMomentOfInertia()
+{
+  const double l = 3, w = 4, h = 5, m = 1.0;
+  const double expectedIx = 3.41667;
+  const double expectedIy = 2.83333;
+  const double expectedIz = 2.08333;
+
+  msgs::Vector3d *size = new msgs::Vector3d();
+  size->set_x(l);
+  size->set_y(w);
+  size->set_z(h);
+
+  msgs::BoxGeom *box = new msgs::BoxGeom();
+  box->set_allocated_size(size);
+
+  msgs::Geometry *geo = new msgs::Geometry();
+  geo->set_type(msgs::Geometry_Type_POLYLINE);
+  geo->set_allocated_box(box);
+
+  msgs::Collision *col = new msgs::Collision();
+  col->set_allocated_geometry(geo);
+
+  ignition::math::Vector3d I = gui::LinkData::ComputeMomentOfInertia(*col, m);
+
+  QVERIFY(fabs(expectedIx - I.X()) < 1e-3);
+  QVERIFY(fabs(expectedIy - I.Y()) < 1e-3);
+  QVERIFY(fabs(expectedIz - I.Z()) < 1e-3);
+
+  delete col;
 }
 
 // Generate a main function for the test